+++ src/cuda_dpm.h
@@ -227,7 +227,7 @@ class dpm {
 
   // cuda setters
   void setDeviceVariables();
-  void setBlockGridDims(int dimBlock);
+  void setBlockGridDims(long dimBlock);
   void cudaVertexNVE(ofstream& enout, double T, double dt0, int NT, int NPRINTSKIP);
 
   // File openers
